Native16BitShaderOpsSupported is false

+++ Sources/Workspace/RenderImage.swift
@@ -108,7 +108,7 @@ func createRenderImage(atoms: [SIMD4<Float>]) -> String {
     // Raster the atoms in order of depth.
     float maximumDepth = -1e38;
     uint32_t hitAtomicNumber = 0;
-    for (uint32_t atomID = 0; atomID < \(atoms.count); ++atomID)
+    for (uint16_t atomID = 0; atomID < \(atoms.count); ++atomID)
     {
       float4 atom = moleculeCoordinates[atomID];
       uint32_t atomicNumber = uint32_t(atom[3]);
